Extract MapSync library from BugReporterMetrics class
